Create commands.json
This is pretty simple to understand but I'll explain it anyway. This JSON file will map where the script is for executing the desired command. So automation:spam:email will lead to the directory where that script is located. This is pretty inefficient compared to my other ideas I've had but it'll do for now.